OPeNDAP: change chunk from {} to null
There's an issue with the interworkings of dask and OPeNDAP: If dask
decides to do chunks on a dataset, it will request the whole chunk from
the backend storage, even if only a part of it is needed in the end.
However for OPeNDAP it is best to request only the required parts from
the backend. This change disables initial dask-chunking for all datasets
which previously used only a single chunk.

While this likely will improve performance for a lot of cases, it may
not always be the best solution. If multiple datasets are to be
concatenated, dask has to be used. However, in those cases, the user
maybe knows about chunking an will define those manually.